+++ tests/misc/programs/misc/src/context.rs
@@ -821,7 +821,7 @@ pub struct Empty {}
 pub struct TestMultipleZeroConstraint<'info> {
     #[account(zero)]
     pub one: Account<'info, Data>,
-    #[account(zero)]
+    #[account(zero, dup)] // Allow duplicate accounts to test the error
     pub two: Account<'info, Data>,
 }
 
@@ -829,7 +829,7 @@ pub struct TestMultipleZeroConstraint<'info> {
 pub struct TestInitAndZero<'info> {
     #[account(init, payer = payer, space = Data::DISCRIMINATOR.len() + Data::LEN)]
     pub init: Account<'info, Data>,
-    #[account(zero)]
+    #[account(zero, dup)] // Allow duplicate accounts to test the error
     pub zero: Account<'info, Data>,
     #[account(mut)]
     pub payer: Signer<'info>,